Separated C++ implementation logic into multiple files based on functionality.
