+++ src/Native/LibTorchSharp/CMakeLists.txt
@@ -16,11 +16,21 @@ set(SOURCES
     THSTorch.h
     Utils.h
     cifar10.cpp
+	THSActivation.cpp
     THSAutograd.cpp
+	THSConvolution.cpp
     THSData.cpp
     THSJIT.cpp
+	THSLoss.cpp
+	THSModule.cpp
     THSNN.cpp
+	THSNormalization.cpp
+	THSOptimizers.cpp
+	THSRandom.cpp
     THSTensor.cpp
+	THSTensorConv.cpp
+	THSTensorFactories.cpp
+	THSTensorMath.cpp
     THSTorch.cpp
     Utils.cpp)
