+++ src/Native/LibTorchSharp/THSTensor.cpp
@@ -4,142 +4,6 @@
 #include <iostream>
 #include <fstream>
 
-Tensor THSTensor_abs(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->abs());
-}
-
-Tensor THSTensor_abs_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->abs_());
-}
-
-Tensor THSTensor_acos(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->acos());
-}
-
-Tensor THSTensor_acos_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->acos_());
-}
-
-Tensor THSTensor_adaptive_avg_pool1d(
-    const Tensor tensor,
-    const int64_t* outputSize, const int outputSizeLength)
-{
-    CATCH_TENSOR(torch::adaptive_avg_pool1d(
-        *tensor,
-        at::ArrayRef<int64_t>(outputSize, outputSizeLength)));
-}
-
-Tensor THSTensor_adaptive_avg_pool2d(
-    const Tensor tensor,
-    const int64_t* outputSize, const int outputSizeLength)
-{
-    CATCH_TENSOR(torch::adaptive_avg_pool2d(
-        *tensor,
-        at::ArrayRef<int64_t>(outputSize, outputSizeLength)));
-}
-
-Tensor THSTensor_adaptive_avg_pool3d(
-    const Tensor tensor,
-    const int64_t* outputSize, const int outputSizeLength)
-{
-    CATCH_TENSOR(torch::adaptive_avg_pool3d(
-        *tensor,
-        at::ArrayRef<int64_t>(outputSize, outputSizeLength)));
-}
-
-Tensor THSTensor_adaptive_avg_pool3d_backward(
-    const Tensor grad_output,
-    const Tensor tensor)
-{
-    CATCH_TENSOR(torch::adaptive_avg_pool3d_backward(
-        *grad_output,
-        *tensor));
-}
-
-Tensor THSTensor_add(const Tensor left, const Tensor right, const Scalar alpha)
-{
-    CATCH_TENSOR(left->add(*right, *alpha));
-}
-
-Tensor THSTensor_add_(const Tensor left, const Tensor right, const Scalar alpha)
-{
-    CATCH_TENSOR(left->add_(*right, *alpha));
-}
-
-Tensor THSTensor_add_scalar(const Tensor left, const Scalar right, const Scalar alpha)
-{
-    CATCH_TENSOR(left->add(*right, *alpha));
-}
-
-Tensor THSTensor_add_scalar_(const Tensor left, const Scalar right, const Scalar alpha)
-{
-    CATCH_TENSOR(left->add_(*right, *alpha));
-}
-
-Tensor THSTensor_addbmm(const Tensor mat, const Tensor batch1, const Tensor batch2, const float beta, const float alpha)
-{
-    CATCH_TENSOR(mat->addbmm(*batch1, *batch2, beta, alpha));
-}
-
-Tensor THSTensor_addbmm_(const Tensor mat, const Tensor batch1, const Tensor batch2, const float beta, const float alpha)
-{
-    CATCH_TENSOR(mat->addbmm_(*batch1, *batch2, beta, alpha));
-}
-
-Tensor THSTensor_addcdiv(const Tensor left, const Tensor tensor1, const Tensor tensor2, const Scalar value)
-{
-    CATCH_TENSOR(left->addcdiv(*tensor1, *tensor2, *value));
-}
-
-Tensor THSTensor_addcdiv_(const Tensor left, const Tensor tensor1, const Tensor tensor2, const Scalar value)
-{
-    CATCH_TENSOR(left->addcdiv_(*tensor1, *tensor2, *value));
-}
-
-Tensor THSTensor_addcmul(const Tensor left, const Tensor tensor1, const Tensor tensor2, const Scalar value)
-{
-    CATCH_TENSOR(left->addcmul(*tensor1, *tensor2, *value));
-}
-
-Tensor THSTensor_addcmul_(const Tensor left, const Tensor tensor1, const Tensor tensor2, const Scalar value)
-{
-    CATCH_TENSOR(left->addcmul_(*tensor1, *tensor2, *value));
-}
-
-Tensor THSTensor_addmm(const Tensor mat, const Tensor mat1, const Tensor mat2, const float beta, const float alpha)
-{
-    CATCH_TENSOR(mat->addmm(*mat1, *mat2, beta, alpha));
-}
-
-Tensor THSTensor_addmm_(const Tensor mat, const Tensor mat1, const Tensor mat2, const float beta, const float alpha)
-{
-    CATCH_TENSOR(mat->addmm_(*mat1, *mat2, beta, alpha));
-}
-
-Tensor THSTensor_addmv(const Tensor mat, const Tensor mat1, const Tensor vec2, const float beta, const float alpha)
-{
-    CATCH_TENSOR(mat->addmv(*mat1, *vec2, beta, alpha));
-}
-
-Tensor THSTensor_addmv_(const Tensor mat, const Tensor mat1, const Tensor vec2, const float beta, const float alpha)
-{
-    CATCH_TENSOR(mat->addmv_(*mat1, *vec2, beta, alpha));
-}
-
-Tensor THSTensor_addr(const Tensor mat, const Tensor mat1, const Tensor vec2, const float beta, const float alpha)
-{
-    CATCH_TENSOR(mat->addr(*mat1, *vec2, beta, alpha));
-}
-
-Tensor THSTensor_addr_(const Tensor mat, const Tensor mat1, const Tensor vec2, const float beta, const float alpha)
-{
-    CATCH_TENSOR(mat->addr_(*mat1, *vec2, beta, alpha));
-}
-
 int THSTensor_allclose(const Tensor left, const Tensor right, double rtol, double atol, bool equal_nan)
 {
     CATCH_RETURN(int, 0, left->allclose(*right, rtol, atol, equal_nan));
@@ -169,74 +33,6 @@ Tensor THSTensor_any_along_dimension(const Tensor tensor, const int64_t dim, boo
 {
     CATCH_TENSOR(tensor->any(dim, keepdim));
 }
-
-Tensor THSTensor_arange(const Scalar start, const Scalar end, const Scalar step, const int8_t scalar_type, const int device_type, const int device_index, const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::arange(*start, *end, *step, options));
-}
-
-Tensor THSTensor_arange_out(const Scalar start, const Scalar end, const Scalar step, const Tensor out)
-{
-    CATCH_TENSOR(torch::arange_out(*out, *start, *end, *step));
-}
-
-
-Tensor THSTensor_linspace(const double start, const double end, const int64_t steps, const int8_t scalar_type, const int device_type, const int device_index, const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::linspace(start, end, steps, options));
-}
-
-Tensor THSTensor_logspace(const double start, const double end, const int64_t steps, double base, const int8_t scalar_type, const int device_type, const int device_index, const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::logspace(start, end, steps, base, options));
-}
-
-
-Tensor THSTensor_arccosh(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->arccosh());
-}
-
-Tensor THSTensor_arccosh_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->arccosh_());
-}
-
-Tensor THSTensor_arcsinh(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->arcsinh());
-}
-
-Tensor THSTensor_arcsinh_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->arcsinh_());
-}
-
-Tensor THSTensor_arctanh(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->arctanh());
-}
-
-Tensor THSTensor_arctanh_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->arctanh_());
-}
-
 Tensor THSTensor_argmax(const Tensor tensor)
 {
     CATCH_TENSOR(tensor->argmax());
@@ -257,36 +53,6 @@ Tensor THSTensor_argmin_along_dimension(const Tensor tensor, const int64_t dim, 
     CATCH_TENSOR(tensor->argmin(dim, keepdim));
 }
 
-Tensor THSTensor_asin(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->asin());
-}
-
-Tensor THSTensor_asin_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->asin_());
-}
-
-Tensor THSTensor_atan(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->atan());
-}
-
-Tensor THSTensor_atan_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->atan_());
-}
-
-Tensor THSTensor_atan2(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->atan2(*other));
-}
-
-Tensor THSTensor_atan2_(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->atan2_(*other));
-}
-
 Tensor THSTensor_atleast_1d(const Tensor tensor)
 {
     CATCH_TENSOR(torch::atleast_1d(*tensor));
@@ -302,99 +68,6 @@ Tensor THSTensor_atleast_3d(const Tensor tensor)
     CATCH_TENSOR(torch::atleast_3d(*tensor));
 }
 
-Tensor THSTensor_avg_pool1d(
-    const Tensor tensor,
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    bool ceil_mode,
-    bool count_include_pad)
-{
-    CATCH_TENSOR(torch::avg_pool1d(
-        *tensor,
-        at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        ceil_mode,
-        count_include_pad));
-}
-
-Tensor THSTensor_avg_pool2d(
-    const Tensor tensor,
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    bool ceil_mode,
-    bool count_include_pad)
-{
-    CATCH_TENSOR(torch::avg_pool2d(
-        *tensor,
-        at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        ceil_mode,
-        count_include_pad));
-}
-
-Tensor THSTensor_avg_pool2d_backward(
-    const Tensor grad_output,
-    const Tensor tensor,
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    bool ceil_mode,
-    bool count_include_pad,
-    const int64_t divisor_override)
-{
-    CATCH_TENSOR(torch::avg_pool2d_backward(
-        *grad_output,
-        *tensor,
-        at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        ceil_mode,
-        count_include_pad,
-        (divisor_override == 0 ? c10::optional<int64_t>() : c10::optional<int64_t>(divisor_override))));
-}
-
-Tensor THSTensor_avg_pool3d(
-    const Tensor tensor,
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    bool ceil_mode,
-    bool count_include_pad)
-{
-    CATCH_TENSOR(torch::avg_pool3d(
-        *tensor,
-        at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        ceil_mode,
-        count_include_pad));
-}
-
-Tensor THSTensor_avg_pool3d_backward(
-    const Tensor grad_output,
-    const Tensor tensor,
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    bool ceil_mode,
-    bool count_include_pad,
-    const int64_t divisor_override)
-{
-    CATCH_TENSOR(torch::avg_pool3d_backward(
-        *grad_output,
-        *tensor,
-        at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        ceil_mode,
-        count_include_pad,
-        (divisor_override == 0 ? c10::optional<int64_t>() : c10::optional<int64_t>(divisor_override))));
-}
-
 void THSTensor_backward(Tensor tensor)
 {
     CATCH(
@@ -402,69 +75,14 @@ void THSTensor_backward(Tensor tensor)
     )
 }
 
-Tensor THSTensor_baddbmm(
-    const Tensor batch1,
-    const Tensor batch2,
-    const Tensor mat,
-    const float beta,
-    const float alpha)
-{
-    CATCH_TENSOR(mat->baddbmm(*batch1, *batch2, beta, alpha));
-}
-
 Tensor THSTensor_bincount(const Tensor tensor, const Tensor weights, const int64_t minlength)
 {
     CATCH_TENSOR(tensor->bincount((weights ? *weights : at::Tensor()), minlength));
 }
 
-Tensor THSTensor_bitwise_and(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->bitwise_and(*other));
-}
-
-Tensor THSTensor_bitwise_and_(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->bitwise_and_(*other));
-}
-
-Tensor THSTensor_bitwise_not(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->bitwise_not());
-}
-
-Tensor THSTensor_bitwise_not_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->bitwise_not_());
-}
-
-Tensor THSTensor_bitwise_or(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->bitwise_or(*other));
-}
-
-Tensor THSTensor_bitwise_or_(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->bitwise_or_(*other));
-}
-
-Tensor THSTensor_bitwise_xor(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->bitwise_xor(*other));
-}
-
-Tensor THSTensor_bitwise_xor_(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->bitwise_xor_(*other));
-}
-
 Tensor THSTensor_block_diag(const Tensor* tensors, const int length)
 {
     CATCH_TENSOR(torch::block_diag(toTensors<at::Tensor>((torch::Tensor**)tensors, length)));
-}
-
-Tensor THSTensor_bmm(const Tensor batch1, const Tensor batch2)
-{
-    CATCH_TENSOR(batch1->bmm(*batch2));
 }
 
 Tensor THSTensor_broadcast_to(const Tensor tensor, const int64_t* shape, const int shape_len)
@@ -491,16 +109,6 @@ Tensor THSTensor_cat(const Tensor* tensors, const int length, const int64_t dim)
     CATCH_TENSOR(torch::cat(toTensors<at::Tensor>((torch::Tensor**)tensors, length), dim));
 }
 
-Tensor THSTensor_ceil(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->ceil());
-}
-
-Tensor THSTensor_ceil_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->ceil_());
-}
-
 Tensor THSTensor_celu(const Tensor tensor)
 {
     CATCH_TENSOR(torch::celu(*tensor));
@@ -524,6 +132,17 @@ Tensor THSTensor_cholesky_inverse(const Tensor tensor, const bool upper)
 Tensor THSTensor_cholesky_solve(const Tensor tensor, const Tensor tensor2, const bool upper)
 {
     CATCH_TENSOR(tensor->cholesky_solve(*tensor2, upper));
+}
+
+void THSTensor_chunk(const Tensor tensor, Tensor* (*allocator)(size_t length), const int64_t chunks, const int64_t dim)
+{
+    CATCH(
+        auto res = tensor->chunk(chunks, dim);
+        const size_t sz = res.size();
+        Tensor * result = allocator(sz);
+        for (size_t i = 0; i < sz; i++)
+            result[i] = new torch::Tensor(res[i]);
+    )
 }
 
 Tensor THSTensor_clamp(const Tensor tensor, const Scalar min, const Scalar max)
@@ -571,130 +190,9 @@ Tensor THSTensor_contiguous(const Tensor tensor)
     CATCH_TENSOR(tensor->contiguous());
 }
 
-Tensor THSTensor_conv_transpose1d(
-    const Tensor input, const Tensor weight, const Tensor bias,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* outputPadding, const int outputPaddingLength,
-    const int64_t* dilation, const int dilationLength,
-    int64_t groups)
-{
-    CATCH_TENSOR(torch::conv_transpose1d(*input, *weight, (bias ? *bias : at::Tensor()),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        at::ArrayRef<int64_t>(outputPadding, outputPaddingLength),
-        groups,
-        at::ArrayRef<int64_t>(dilation, dilationLength)));
-}
-
-Tensor THSTensor_conv_transpose2d(
-    const Tensor input, const Tensor weight, const Tensor bias,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* outputPadding, const int outputPaddingLength,
-    const int64_t* dilation, const int dilationLength,
-    int64_t groups)
-{
-    CATCH_TENSOR(torch::conv_transpose2d(*input, *weight, (bias ? *bias : at::Tensor()),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        at::ArrayRef<int64_t>(outputPadding, outputPaddingLength),
-        groups,
-        at::ArrayRef<int64_t>(dilation, dilationLength)));
-}
-
-Tensor THSTensor_conv_transpose3d(
-    const Tensor input, const Tensor weight, const Tensor bias,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* outputPadding, const int outputPaddingLength,
-    const int64_t* dilation, const int dilationLength,
-    int64_t groups)
-{
-    CATCH_TENSOR(torch::conv_transpose3d(*input, *weight, (bias ? *bias : at::Tensor()),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        at::ArrayRef<int64_t>(outputPadding, outputPaddingLength),
-        groups,
-        at::ArrayRef<int64_t>(dilation, dilationLength)));
-}
-
-Tensor THSTensor_conv1d(
-    const Tensor input,
-    const Tensor weight,
-    const Tensor bias,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* dilation, const int dilationLength,
-    int64_t groups)
-{
-    CATCH_TENSOR(torch::conv1d(*input, *weight, (bias ? *bias : at::Tensor()),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        at::ArrayRef<int64_t>(dilation, dilationLength),
-        groups));
-}
-
-Tensor THSTensor_conv2d(
-    const Tensor input,
-    const Tensor weight,
-    const Tensor bias,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* dilation, const int dilationLength,
-    int64_t groups)
-{
-    CATCH_TENSOR(torch::conv2d(*input, *weight, (bias ? *bias : at::Tensor()),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        at::ArrayRef<int64_t>(dilation, dilationLength),
-        groups));
-}
-
-Tensor THSTensor_conv3d(
-    const Tensor input,
-    const Tensor weight,
-    const Tensor bias,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* dilation, const int dilationLength,
-    int64_t groups)
-{
-    CATCH_TENSOR(torch::conv3d(*input, *weight, (bias ? *bias : at::Tensor()),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        at::ArrayRef<int64_t>(dilation, dilationLength),
-        groups));
-}
-
 Tensor THSTensor_copysign(const Tensor input, const Tensor other)
 {
     CATCH_TENSOR(input->copysign(*other));
-}
-
-Tensor THSTensor_cos(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->cos());
-}
-
-Tensor THSTensor_cos_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->cos_());
-}
-
-Tensor THSTensor_cosh(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->cosh());
-}
-
-Tensor THSTensor_cosh_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->cosh_());
-}
-
-Tensor THSTensor_cross(const Tensor tensor, const Tensor other, const int64_t dim)
-{
-    CATCH_TENSOR(tensor->cross(*other, dim));
 }
 
 Tensor THSTensor_cpu(const Tensor tensor)
@@ -746,11 +244,6 @@ float THSTensor_data_idx_float16(const Tensor tensor, const int64_t i)
 float THSTensor_data_idx_bfloat16(const Tensor tensor, const int64_t i)
 {
     CATCH_RETURN(float, NULL, (float)(tensor->data_ptr<c10::BFloat16>())[i]);
-}
-
-Tensor THSTensor_deg2rad(const Tensor tensor)
-{
-    CATCH_TENSOR(torch::deg2rad(*tensor));
 }
 
 const char* THSTensor_device_str(const Tensor tensor)
@@ -814,36 +307,6 @@ Tensor THSTensor_dist(const Tensor tensor, const Tensor other, const float p)
     CATCH_TENSOR(tensor->dist(*other, p));
 }
 
-Tensor THSTensor_div(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->div(*right));
-}
-
-Tensor THSTensor_div_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->div_(*right));
-}
-
-Tensor THSTensor_div_scalar(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->div(*right));
-}
-
-Tensor THSTensor_div_scalar_(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->div_(*right));
-}
-
-Tensor THSTensor_dot(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->dot(*right));
-}
-
-Tensor THSTensor_einsum(const char* equation, const Tensor* tensors, const int length)
-{
-    CATCH_TENSOR(torch::einsum(equation, toTensors<at::Tensor>((torch::Tensor**)tensors, length)));
-}
-
 int64_t THSTensor_element_size(const Tensor tensor)
 {
     CATCH_RETURN(int64_t, 0, tensor->element_size());
@@ -857,164 +320,6 @@ Tensor THSTensor_elu(const Tensor tensor, const Scalar alpha, const Scalar scale
 Tensor THSTensor_elu_(const Tensor tensor, const Scalar alpha, const Scalar scale, const Scalar input_scale)
 {
     CATCH_TENSOR(torch::elu_(*tensor, *alpha, *scale, *input_scale));
-}
-
-Tensor THSTensor_empty_out(const int64_t* sizes, const int length, const Tensor out)
-{
-    CATCH_TENSOR(torch::empty_out(*out, at::ArrayRef<int64_t>(sizes, length)));
-}
-
-Tensor THSTensor_empty(
-    const int64_t* sizes,
-    const int length,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::empty(at::ArrayRef<int64_t>(sizes, length), options));
-}
-
-Tensor THSTensor_empty_like(
-    const Tensor input,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::empty_like(*input, options));
-}
-
-Tensor THSTensor_empty_strided(
-    const int64_t* sizes,
-    const int sz_length,
-    const int64_t* strides,
-    const int str_length,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::empty_strided(at::ArrayRef<int64_t>(sizes, sz_length), at::ArrayRef<int64_t>(strides, str_length), options));
-}
-
-Tensor THSTensor_as_strided(
-    const Tensor input,
-    const int64_t* sizes,
-    const int sz_length,
-    const int64_t* strides,
-    const int str_length,
-    const int64_t storage_offset)
-{
-    CATCH_TENSOR(input->as_strided(at::ArrayRef<int64_t>(sizes, sz_length), at::ArrayRef<int64_t>(strides, str_length), storage_offset));
-}
-
-Tensor THSTensor_eye(const int64_t n, const int64_t m, const int8_t scalar_type, const int device_type, const int device_index, const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::eye(n, m, options));
-}
-
-Tensor THSTensor_eye_out(const int64_t n, const int64_t m, const Tensor out)
-{
-    CATCH_TENSOR(torch::eye_out(*out, n, m));
-}
-
-
-Tensor THSTensor_eq(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->eq(*right));
-}
-
-Tensor THSTensor_eq_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->eq_(*right));
-}
-
-Tensor THSTensor_eq_scalar(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->eq(*right));
-}
-
-Tensor THSTensor_eq_scalar_(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->eq_(*right));
-}
-
-int THSTensor_equal(const Tensor left, const Tensor right)
-{
-    CATCH_RETURN(int, 0, left->equal(*right));
-}
-
-Tensor THSTensor_exp(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->exp());
-}
-
-Tensor THSTensor_exp2(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->exp2());
-}
-
-Tensor THSTensor_exp_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->exp_());
-}
-
-Tensor THSTensor_expm1(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->expm1());
-}
-
-Tensor THSTensor_expm1_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->expm1_());
-}
-
-Tensor THSTensor_erf(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->erf());
-}
-
-Tensor THSTensor_erf_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->erf_());
-}
-
-Tensor THSTensor_erfc(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->erfc());
-}
-
-Tensor THSTensor_erfc_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->erfc_());
-}
-
-Tensor THSTensor_erfinv(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->erfinv());
-}
-
-Tensor THSTensor_erfinv_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->erfinv_());
 }
 
 Tensor THSTensor_expand(const Tensor tensor, const int64_t* sizes, const int length, bool implicit)
@@ -1086,87 +391,12 @@ Tensor THSTensor_fill_(const Tensor tensor, const Scalar value)
     CATCH_TENSOR(tensor->fill_(*value));
 }
 
-Tensor THSTensor_float_power(const Tensor input, const Tensor exponent)
-{
-    CATCH_TENSOR(input->float_power(*exponent));
-}
-
-Tensor THSTensor_floor(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->floor());
-}
-
-Tensor THSTensor_floor_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->floor_());
-}
-
-Tensor THSTensor_fmax(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->fmax(*right));
-}
-
-Tensor THSTensor_fmin(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->fmin(*right));
-}
-
-Tensor THSTensor_fmod(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->fmod(*right));
-}
-
-Tensor THSTensor_fmod_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->fmod_(*right));
-}
-
-Tensor THSTensor_fmod_scalar(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->fmod(*right));
-}
-
-Tensor THSTensor_fmod_scalar_(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->fmod_(*right));
-}
-
-Tensor THSTensor_frac(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->frac());
-}
-
-Tensor THSTensor_frac_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->frac_());
-}
-
 Tensor THSTensor_gather(
     const Tensor tensor,
     const int64_t dim,
     const Tensor index)
 {
     CATCH_TENSOR(torch::gather(*tensor, dim, *index));
-}
-
-Tensor THSTensor_ge(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->ge(*right));
-}
-
-Tensor THSTensor_ge_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->ge_(*right));
-}
-
-Tensor THSTensor_ge_scalar(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->ge(*right));
-}
-
-Tensor THSTensor_ge_scalar_(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->ge_(*right));
 }
 
 Tensor THSTensor_gelu(const Tensor tensor)
@@ -1222,26 +452,6 @@ Tensor THSTensor_grad(const Tensor tensor)
     res = grad.defined() ? new torch::Tensor(grad) : NULL;
     );
     return res;
-}
-
-Tensor THSTensor_gt(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->gt(*right));
-}
-
-Tensor THSTensor_gt_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->gt_(*right));
-}
-
-Tensor THSTensor_gt_scalar(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->gt(*right));
-}
-
-Tensor THSTensor_gt_scalar_(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->gt_(*right));
 }
 
 Tensor THSTensor_hardsigmoid(const Tensor tensor)
@@ -1489,31 +699,6 @@ Tensor THSTensor_lcm_(const Tensor tensor, const Tensor other)
     CATCH_TENSOR(tensor->lcm_(*other));
 }
 
-Tensor THSTensor_ldexp(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->ldexp(*right));
-}
-
-Tensor THSTensor_le(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->le(*right));
-}
-
-Tensor THSTensor_le_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->le_(*right));
-}
-
-Tensor THSTensor_le_scalar(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->le(*right));
-}
-
-Tensor THSTensor_le_scalar_(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->le_(*right));
-}
-
 Tensor THSTensor_leaky_relu(const Tensor tensor, const Scalar negative_slope)
 {
     CATCH_TENSOR(torch::leaky_relu(*tensor, *negative_slope));
@@ -1533,8 +718,8 @@ Tensor THSTensor_load(const char* location)
 {
     CATCH_RETURN_Tensor(
         torch::Tensor tensor;
-    torch::load(tensor, location);
-    res = ResultTensor(tensor);
+        torch::load(tensor, location);
+        res = ResultTensor(tensor);
     );
 }
 
@@ -1543,71 +728,10 @@ Tensor THSTensor_log_sigmoid(const Tensor tensor)
     CATCH_TENSOR(torch::log_sigmoid(*tensor));
 }
 
-Tensor THSTensor_logcumsumexp(const Tensor tensor, const long dimension)
-{
-    CATCH_TENSOR(torch::logcumsumexp(*tensor, dimension));
-}
-
-Tensor THSTensor_logaddexp(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(torch::logaddexp(*tensor, *other));
-}
-
-Tensor THSTensor_logaddexp2(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(torch::logaddexp2(*tensor, *other));
-}
-
-Tensor THSTensor_logsumexp(const Tensor tensor, const long dim, const bool keepdim)
-{
-    CATCH_TENSOR(torch::logsumexp(*tensor, dim, keepdim));
-}
-
-
 //Tensor THSTensor_log_sigmoid_backward(const Tensor tensor)
 //{
 //    CATCH_TENSOR(torch::log_sigmoid_backward(*tensor));
 //}
-
-Tensor THSTensor_log(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->log());
-}
-
-Tensor THSTensor_log_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->log_());
-}
-
-Tensor THSTensor_log2(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->log2());
-}
-
-Tensor THSTensor_log2_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->log2_());
-}
-
-Tensor THSTensor_log10(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->log10());
-}
-
-Tensor THSTensor_log10_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->log10_());
-}
-
-Tensor THSTensor_log1p(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->log1p());
-}
-
-Tensor THSTensor_log1p_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->log1p_());
-}
 
 Tensor THSTensor_lerp(const Tensor tensor, const Tensor end, const Tensor weight)
 {
@@ -1617,71 +741,6 @@ Tensor THSTensor_lerp(const Tensor tensor, const Tensor end, const Tensor weight
 Tensor THSTensor_lerp_(const Tensor tensor, const Tensor end, const Tensor weight)
 {
     CATCH_TENSOR(tensor->lerp_(*end, *weight));
-}
-
-Tensor THSTensor_logical_and(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->logical_and(*other));
-}
-
-Tensor THSTensor_logical_and_(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->logical_and_(*other));
-}
-
-Tensor THSTensor_logical_not(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->logical_not());
-}
-
-Tensor THSTensor_logical_not_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->logical_not_());
-}
-
-Tensor THSTensor_logical_or(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->logical_or(*other));
-}
-
-Tensor THSTensor_logical_or_(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->logical_or_(*other));
-}
-
-Tensor THSTensor_logical_xor(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->logical_xor(*other));
-}
-
-Tensor THSTensor_logical_xor_(const Tensor tensor, const Tensor other)
-{
-    CATCH_TENSOR(tensor->logical_xor_(*other));
-}
-
-Tensor THSTensor_logit(const Tensor tensor, const double* eps)
-{
-    CATCH_TENSOR((eps == nullptr) ? tensor->logit() : tensor->logit(*eps));
-}
-
-Tensor THSTensor_lt(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->lt(*right));
-}
-
-Tensor THSTensor_lt_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->lt_(*right));
-}
-
-Tensor THSTensor_lt_scalar(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->lt(*right));
-}
-
-Tensor THSTensor_lt_scalar_(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->lt_(*right));
 }
 
 Tensor THSTensor_masked_fill(const Tensor tensor, const Tensor mask, const Scalar value)
@@ -1709,11 +768,6 @@ Tensor THSTensor_masked_select(const Tensor tensor, const Tensor mask)
     CATCH_TENSOR(tensor->masked_select(*mask));
 }
 
-Tensor THSTensor_matmul(const Tensor left, const Tensor right)
-{
-    return  new torch::Tensor(left->matmul(*right));
-}
-
 Tensor THSTensor_max(const Tensor tensor)
 {
     CATCH_TENSOR(tensor->max());
@@ -1733,158 +787,10 @@ void THSTensor_max_along_dimension(const Tensor tensor, Tensor* (*allocator)(siz
 {
     CATCH(
         auto max = tensor->max(dim, keepdim);
-    Tensor * result = allocator(2);
-    result[0] = new torch::Tensor(std::get<0>(max));
-    result[1] = new torch::Tensor(std::get<1>(max));
+        Tensor * result = allocator(2);
+        result[0] = new torch::Tensor(std::get<0>(max));
+        result[1] = new torch::Tensor(std::get<1>(max));
     )
-}
-
-
-Tensor THSTensor_max_pool1d(
-    const Tensor tensor,
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* dilation, const int dilationLength,
-    bool ceil_mode)
-{
-    CATCH_TENSOR(torch::max_pool1d(
-        *tensor,
-        at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        at::ArrayRef<int64_t>(dilation, dilationLength),
-        ceil_mode));
-}
-
-void THSTensor_max_pool1d_with_indices(
-    const Tensor tensor,
-    Tensor* (*allocator)(size_t length),
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* dilation, const int dilationLength,
-    bool ceil_mode)
-{
-    CATCH(
-        auto res = torch::max_pool1d_with_indices(
-            *tensor,
-            at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-            at::ArrayRef<int64_t>(stride, strideLength),
-            at::ArrayRef<int64_t>(padding, paddingLength),
-            at::ArrayRef<int64_t>(dilation, dilationLength),
-            ceil_mode);
-
-    Tensor * result = allocator(2);
-    result[0] = new torch::Tensor(std::get<0>(res));
-    result[1] = new torch::Tensor(std::get<1>(res));
-    )
-}
-
-Tensor THSTensor_max_pool2d(
-    const Tensor tensor,
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* dilation, const int dilationLength,
-    bool ceil_mode)
-{
-    CATCH_TENSOR(torch::max_pool2d(
-        *tensor,
-        at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        at::ArrayRef<int64_t>(dilation, dilationLength),
-        ceil_mode));
-}
-
-void THSTensor_max_pool2d_with_indices(
-    const Tensor tensor,
-    Tensor* (*allocator)(size_t length),
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* dilation, const int dilationLength,
-    bool ceil_mode)
-{
-    CATCH(
-        auto res = torch::max_pool2d_with_indices(
-            *tensor,
-            at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-            at::ArrayRef<int64_t>(stride, strideLength),
-            at::ArrayRef<int64_t>(padding, paddingLength),
-            at::ArrayRef<int64_t>(dilation, dilationLength),
-            ceil_mode);
-    Tensor * result = allocator(2);
-    result[0] = new torch::Tensor(std::get<0>(res));
-    result[1] = new torch::Tensor(std::get<1>(res));
-    )
-}
-
-Tensor THSTensor_max_pool3d(
-    const Tensor tensor,
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* dilation, const int dilationLength,
-    bool ceil_mode)
-{
-    CATCH_TENSOR(torch::max_pool3d(
-        *tensor,
-        at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength),
-        at::ArrayRef<int64_t>(dilation, dilationLength),
-        ceil_mode));
-}
-
-void THSTensor_max_pool3d_with_indices(
-    const Tensor tensor,
-    Tensor* (*allocator)(size_t length),
-    const int64_t* kernelSize, const int kernelSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength,
-    const int64_t* dilation, const int dilationLength,
-    bool ceil_mode)
-{
-    CATCH(
-        auto res = torch::max_pool3d_with_indices(
-            *tensor,
-            at::ArrayRef<int64_t>(kernelSize, kernelSizeLength),
-            at::ArrayRef<int64_t>(stride, strideLength),
-            at::ArrayRef<int64_t>(padding, paddingLength),
-            at::ArrayRef<int64_t>(dilation, dilationLength),
-            ceil_mode);
-    Tensor * result = allocator(2);
-    result[0] = new torch::Tensor(std::get<0>(res));
-    result[1] = new torch::Tensor(std::get<1>(res));
-    )
-}
-
-Tensor THSTensor_maxunpool2d(
-    const Tensor tensor,
-    const Tensor indices,
-    const int64_t* outputSize, const int outputSizeLength)
-{
-    CATCH_TENSOR(torch::max_unpool2d(
-        *tensor,
-        *indices,
-        at::ArrayRef<int64_t>(outputSize, outputSizeLength)));
-}
-
-Tensor THSTensor_maxunpool3d(
-    const Tensor tensor,
-    const Tensor indices,
-    const int64_t* outputSize, const int outputSizeLength,
-    const int64_t* stride, const int strideLength,
-    const int64_t* padding, const int paddingLength)
-{
-    CATCH_TENSOR(torch::max_unpool3d(
-        *tensor,
-        *indices,
-        at::ArrayRef<int64_t>(outputSize, outputSizeLength),
-        at::ArrayRef<int64_t>(stride, strideLength),
-        at::ArrayRef<int64_t>(padding, paddingLength)));
 }
 
 Tensor THSTensor_mean(const Tensor tensor)
@@ -1915,17 +821,6 @@ void THSTensor_mode(const Tensor tensor, Tensor* (*allocator)(size_t length), co
         result[0] = new torch::Tensor(std::get<0>(res));
         result[1] = new torch::Tensor(std::get<1>(res));
     )
-}
-
-void THSTensor_chunk(const Tensor tensor, Tensor* (*allocator)(size_t length), const int64_t chunks, const int64_t dim)
-{
-    CATCH(
-        auto res = tensor->chunk(chunks, dim);
-        const size_t sz = res.size();
-        Tensor* result = allocator(sz);
-        for (size_t i = 0; i < sz; i++)
-            result[i] = new torch::Tensor(res[i]);
-        )
 }
 
 Tensor THSTensor_quantile(const Tensor tensor, const Tensor q, const int64_t dim, const bool keep_dim)
@@ -1968,54 +863,9 @@ void THSTensor_min_along_dimension(const Tensor tensor, Tensor* (*allocator)(siz
 //    CATCH_TENSOR(tensor->median(dim, keepdim));
 //}
 
-Tensor THSTensor_mm(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->mm(*right));
-}
-
-Tensor THSTensor_mv(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->mv(*right));
-}
-
-Tensor THSTensor_vdot(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->vdot(*right));
-}
-
 Tensor THSTensor_msort(const Tensor tensor)
 {
     CATCH_TENSOR(tensor->msort());
-}
-
-Tensor THSTensor_mul(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->mul(*right));
-}
-
-Tensor THSTensor_mul_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->mul_(*right));
-}
-
-Tensor THSTensor_mul_scalar(const Tensor tensor, const Scalar scalar)
-{
-    CATCH_TENSOR(tensor->mul(*scalar));
-}
-
-Tensor THSTensor_mul_scalar_(const Tensor tensor, const Scalar scalar)
-{
-    CATCH_TENSOR(tensor->mul_(*scalar));
-}
-
-Tensor THSTensor_mvlgamma(const Tensor tensor, int64_t p)
-{
-    CATCH_TENSOR(tensor->mvlgamma(p));
-}
-
-Tensor THSTensor_mvlgamma_(const Tensor tensor, int64_t p)
-{
-    CATCH_TENSOR(tensor->mvlgamma_(p));
 }
 
 Tensor THSTensor_nansum(const Tensor input)
@@ -2036,203 +886,9 @@ Tensor THSTensor_nan_to_num(const Tensor input, double* _nan, double* _posinf, d
     CATCH_TENSOR(torch::nan_to_num(*input, nan, posinf, neginf));
 }
 
-Tensor THSTensor_ne(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->ne(*right));
-}
-
 Tensor THSTensor_narrow(const Tensor tensor, int64_t dim, int64_t start, int64_t length)
 {
     CATCH_TENSOR(tensor->narrow(dim, start, length))
-}
-
-Tensor THSTensor_ne_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->ne_(*right));
-}
-
-Tensor THSTensor_ne_scalar(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->ne(*right));
-}
-
-Tensor THSTensor_ne_scalar_(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->ne_(*right));
-}
-
-Tensor THSTensor_neg(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->neg());
-}
-
-Tensor THSTensor_neg_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->neg_());
-}
-
-Tensor THSTensor_new(
-    void* data,
-    void (*deleter)(void*),
-    const int64_t* sizes,
-    const int szlength,
-    int8_t scalar_type,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::from_blob(data, at::ArrayRef<int64_t>(sizes, szlength), deleter, options));
-}
-
-Tensor THSTensor_newInt64(
-    int64_t* data,
-    void (*deleter)(void*),
-    const int64_t* sizes,
-    const int szlength,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(at::kLong))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::from_blob(data, at::ArrayRef<int64_t>(sizes, szlength), deleter, options));
-}
-
-// The data is passed in as float and copied into the array of Half in the C++ code
-// since .NET doesn't know about 'float16' values.
-Tensor THSTensor_newFloat16(
-    float* rawArray,
-    c10::Half* dataArray,
-    void (*deleter)(void*),
-    const int64_t* sizes,
-    const int szlength,
-    const bool requires_grad)
-{
-    CATCH_RETURN_Tensor(
-        int64_t sz = 1;
-    for (int k = 0; k < szlength; k++)
-        sz *= sizes[k];
-    for (int64_t i = 0; i < sz; i++)
-        dataArray[i] = (c10::Half)rawArray[i];
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(at::kHalf))
-        .requires_grad(requires_grad);
-    res = ResultTensor(torch::from_blob(dataArray, at::ArrayRef<int64_t>(sizes, szlength), deleter, options));
-    )
-}
-
-// The data is passed in as float and copied into the array of Half in the C++ code
-// since .NET doesn't know about 'float16' values.
-Tensor THSTensor_newBFloat16(
-    float* rawArray,
-    c10::BFloat16* dataArray,
-    void (*deleter)(void*),
-    const int64_t* sizes,
-    const int szlength,
-    const bool requires_grad)
-{
-    CATCH_RETURN_Tensor(
-        int64_t sz = 1;
-    for (int k = 0; k < szlength; k++)
-        sz *= sizes[k];
-    for (int64_t i = 0; i < sz; i++)
-        dataArray[i] = (c10::BFloat16)rawArray[i];
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(at::kBFloat16))
-        .requires_grad(requires_grad);
-    res = ResultTensor(torch::from_blob(dataArray, at::ArrayRef<int64_t>(sizes, szlength), deleter, options));
-    )
-}
-
-Tensor THSTensor_newInt8Scalar(int8_t data, const int device_type, const int device_index, bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(c10::ScalarType::Char))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::tensor(data, options));
-}
-
-Tensor THSTensor_newByteScalar(char data, const int device_type, const int device_index, bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(c10::ScalarType::Byte))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::tensor(data, options));
-}
-
-Tensor THSTensor_newBoolScalar(bool  data, const int device_type, const int device_index, bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(c10::ScalarType::Bool))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::tensor(data, options));
-}
-
-Tensor THSTensor_newInt16Scalar(short data, const int device_type, const int device_index, bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(c10::ScalarType::Short))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::tensor(data, options));
-}
-
-Tensor THSTensor_newInt32Scalar(int data, const int device_type, const int device_index, bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(c10::ScalarType::Int))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::tensor(data, options));
-}
-
-Tensor THSTensor_newInt64Scalar(int64_t data, const int device_type, const int device_index, bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(c10::ScalarType::Long))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::tensor(data, options));
-}
-
-Tensor THSTensor_newFloat64Scalar(double data, const int device_type, const int device_index, bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(c10::ScalarType::Double))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::tensor(data, options));
-}
-
-Tensor THSTensor_newFloat32Scalar(float data, const int device_type, const int device_index, bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(c10::ScalarType::Float))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::tensor(data, options));
-}
-
-Tensor THSTensor_newFloat16Scalar(float data, const int device_type, const int device_index, bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(c10::ScalarType::Half))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::tensor((c10::Half)data, options));
-}
-
-Tensor THSTensor_newBFloat16Scalar(float data, const int device_type, const int device_index, bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(c10::ScalarType::BFloat16))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-    CATCH_TENSOR(torch::tensor((c10::BFloat16)data, options));
 }
 
 Tensor THSTensor_nextafter(const Tensor input, const Tensor other)
@@ -2260,41 +916,6 @@ int64_t THSTensor_numel(const Tensor tensor)
     CATCH_RETURN(int64_t, 0, tensor->numel());
 }
 
-Tensor THSTensor_ones(
-    const int64_t* sizes,
-    const int length,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::ones(at::ArrayRef<int64_t>(sizes, length), options));
-}
-
-
-Tensor THSTensor_ones_out(const int64_t* sizes, const int length, const Tensor out)
-{
-    CATCH_TENSOR(torch::ones_out(*out, at::ArrayRef<int64_t>(sizes, length)));
-}
-
-Tensor THSTensor_ones_like(
-    const Tensor input,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::ones_like(*input, options));
-}
-
 
 Tensor THSTensor_outer(const Tensor left, const Tensor right)
 {
@@ -2317,26 +938,6 @@ Tensor THSTensor_polygamma_(const Tensor tensor, int64_t n)
     CATCH_TENSOR(tensor->polygamma_(n));
 }
 
-Tensor THSTensor_pow(const Tensor tensor, const Tensor exponent)
-{
-    CATCH_TENSOR(tensor->pow(*exponent));
-}
-
-Tensor THSTensor_pow_(const Tensor tensor, const Tensor exponent)
-{
-    CATCH_TENSOR(tensor->pow_(*exponent));
-}
-
-Tensor THSTensor_pow_scalar(const Tensor tensor, const Scalar exponent)
-{
-    CATCH_TENSOR(tensor->pow(*exponent));
-}
-
-Tensor THSTensor_pow_scalar_(const Tensor tensor, const Scalar exponent)
-{
-    CATCH_TENSOR(tensor->pow_(*exponent));
-}
-
 Tensor THSTensor_prelu(const Tensor left, const Tensor right)
 {
     CATCH_TENSOR(left->prelu(*right));
@@ -2346,149 +947,6 @@ Tensor THSTensor_prelu(const Tensor left, const Tensor right)
 //{
 //    CATCH_TENSOR(torch::prelu_backward(*grad_output, *self, *weight));
 //}
-
-Tensor THSTensor_rad2deg(const Tensor tensor)
-{
-    CATCH_TENSOR(torch::rad2deg(*tensor));
-}
-
-
-Tensor THSTensor_rand(
-    const int64_t* sizes,
-    const int length,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    Tensor tensor;
-    CATCH(
-        auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    tensor = new torch::Tensor(torch::rand(at::ArrayRef<int64_t>(sizes, length), options));
-    )
-        return tensor;
-}
-
-Tensor THSTensor_rand_out(const int64_t* sizes, const int length, const Tensor out)
-{
-    CATCH_TENSOR(torch::rand_out(*out, at::ArrayRef<int64_t>(sizes, length)));
-}
-
-Tensor THSTensor_rand_like(
-    const Tensor input,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::rand_like(*input, options));
-}
-
-Tensor THSTensor_randint(
-    const int64_t high,
-    const int64_t* sizes,
-    const int length,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    Tensor tensor;
-    CATCH(
-        auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    tensor = new torch::Tensor(torch::randint(high, at::ArrayRef<int64_t>(sizes, length), options));
-    )
-        return tensor;
-}
-
-Tensor THSTensor_randint_out(const int64_t high, const int64_t* sizes, const int length, const Tensor out)
-{
-    CATCH_TENSOR(torch::randint_out(*out, high, at::ArrayRef<int64_t>(sizes, length)));
-}
-
-Tensor THSTensor_randint_like(
-    const Tensor input,
-    const int64_t low,
-    const int64_t high,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::randint_like(*input, low, high, options));
-}
-
-Tensor THSTensor_randn(
-    const int64_t* sizes,
-    const int length,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::randn(at::ArrayRef<int64_t>(sizes, length), options));
-}
-
-Tensor THSTensor_randn_out(const int64_t* sizes, const int length, const Tensor out)
-{
-    CATCH_TENSOR(torch::randn_out(*out, at::ArrayRef<int64_t>(sizes, length)));
-}
-
-Tensor THSTensor_randn_like(
-    const Tensor input,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::randn_like(*input, options));
-}
-
-
-Tensor THSTensor_randperm(const int64_t n,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    Tensor tensor;
-    CATCH(
-        auto options =
-        at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    tensor = new torch::Tensor(torch::randperm(n, options));
-    )
-        return tensor;
-}
-
-Tensor THSTensor_randperm_out(const int64_t n, const Tensor out)
-{
-    CATCH_TENSOR(torch::randperm_out(*out, n));
-}
 
 Tensor THSTensor_ravel(const Tensor tensor)
 {
@@ -2505,16 +963,6 @@ Tensor THSTensor_relu_(const Tensor tensor)
     CATCH_TENSOR(torch::relu_(*tensor));
 }
 
-Tensor THSTensor_reciprocal(const Tensor tensor)
-{
-    CATCH_TENSOR(torch::reciprocal(*tensor));
-}
-
-Tensor THSTensor_reciprocal_(const Tensor tensor)
-{
-    CATCH_TENSOR(torch::reciprocal_(*tensor));
-}
-
 Tensor THSTensor_relu6(const Tensor tensor)
 {
     CATCH_TENSOR(torch::nn::functional::relu6(*tensor));
@@ -2523,26 +971,6 @@ Tensor THSTensor_relu6(const Tensor tensor)
 Tensor THSTensor_relu6_(const Tensor tensor)
 {
     CATCH_TENSOR(torch::nn::functional::relu6(*tensor, torch::nn::functional::ReLU6FuncOptions().inplace(true)));
-}
-
-Tensor THSTensor_remainder(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->remainder(*right));
-}
-
-Tensor THSTensor_remainder_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->remainder_(*right));
-}
-
-Tensor THSTensor_remainder_scalar(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->remainder(*right));
-}
-
-Tensor THSTensor_remainder_scalar_(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->remainder_(*right));
 }
 
 Tensor THSTensor_renorm(const Tensor tensor, const float p, const int64_t dim, const float maxnorm)
@@ -2560,31 +988,20 @@ Tensor THSTensor_reshape(const Tensor tensor, const int64_t* shape, const int le
     CATCH_TENSOR(tensor->reshape(at::ArrayRef<int64_t>(shape, length)));
 }
 
-Tensor THSTensor_round(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->round());
-}
-
-Tensor THSTensor_round_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->round_());
-}
-
-Tensor THSTensor_rsqrt(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->rsqrt());
-}
-
-Tensor THSTensor_rsqrt_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->rsqrt_());
-}
-
 void THSTensor_save(const Tensor tensor, const char* location)
 {
     CATCH(
         torch::save(*tensor, location);
     );
+}
+
+Tensor THSTensor_scatter(
+    const Tensor tensor,
+    const int64_t dim,
+    const Tensor index,
+    const Tensor source)
+{
+    CATCH_TENSOR(torch::scatter(*tensor, dim, *index, *source));
 }
 
 Tensor THSTensor_selu(const Tensor tensor)
@@ -2649,48 +1066,14 @@ void THSTensor_set6(const Tensor tensor, int64_t index1, int64_t index2, int64_t
     )
 }
 
-Tensor THSTensor_sign(const Tensor tensor)
+Tensor THSTensor_slice(const Tensor tensor, int64_t dim, int64_t start, int64_t finish, int64_t step)
 {
-    CATCH_TENSOR(tensor->sign());
-}
-
-Tensor THSTensor_sign_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->sign_());
-}
-
-Tensor THSTensor_signbit(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->signbit());
-}
-
-Tensor THSTensor_sqrt(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->sqrt());
-}
-
-Tensor THSTensor_sqrt_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->sqrt_());
+    CATCH_TENSOR(tensor->slice(dim, start, finish, step));
 }
 
 Tensor THSTensor_softplus(const Tensor tensor)
 {
     CATCH_TENSOR(torch::softplus(*tensor));
-}
-
-Tensor THSTensor_scatter(
-    const Tensor tensor,
-    const int64_t dim,
-    const Tensor index,
-    const Tensor source)
-{
-    CATCH_TENSOR(torch::scatter(*tensor, dim, *index, *source));
-}
-
-Tensor THSTensor_slice(const Tensor tensor, int64_t dim, int64_t start, int64_t finish, int64_t step)
-{
-    CATCH_TENSOR(tensor->slice(dim, start, finish, step));
 }
 
 Tensor THSTensor_sparse(
@@ -2718,36 +1101,6 @@ Tensor THSTensor_sigmoid(const Tensor tensor)
 Tensor THSTensor_sigmoid_(const Tensor tensor)
 {
     CATCH_TENSOR(tensor->sigmoid_());
-}
-
-Tensor THSTensor_sin(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->sin());
-}
-
-Tensor THSTensor_sin_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->sin_());
-}
-
-Tensor THSTensor_sinc(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->sinc());
-}
-
-Tensor THSTensor_sinc_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->sinc_());
-}
-
-Tensor THSTensor_sinh(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->sinh());
-}
-
-Tensor THSTensor_sinh_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->sinh_());
 }
 
 void THSTensor_split_with_size(
@@ -2842,16 +1195,6 @@ int64_t THSTensor_size(const Tensor tensor, const int64_t dim)
     CATCH_RETURN(int64_t, 0, tensor->size(dim));
 }
 
-Tensor THSTensor_sub(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->sub(*right));
-}
-
-Tensor THSTensor_sub_(const Tensor left, const Tensor right)
-{
-    CATCH_TENSOR(left->sub_(*right));
-}
-
 Tensor THSTensor_sub_scalar(const Tensor left, const Scalar right)
 {
     CATCH_TENSOR(left->sub(*right));
@@ -2902,11 +1245,6 @@ Tensor THSTensor_std_along_dimensions(const Tensor tensor, const int64_t* dimens
     CATCH_TENSOR(tensor->std(at::ArrayRef<int64_t>(dimensions, length), unbiased, keepdim));
 }
 
-Tensor THSTensor_sub_scalar_(const Tensor left, const Scalar right)
-{
-    CATCH_TENSOR(left->sub_(*right));
-}
-
 Tensor THSTensor_sum(const Tensor tensor, bool has_type, const int8_t dtype)
 {
     CATCH_TENSOR(has_type ? tensor->sum((c10::ScalarType)dtype) : tensor->sum())
@@ -2931,26 +1269,6 @@ Tensor THSTensor_take(const Tensor tensor, const Tensor indices)
     CATCH_TENSOR(tensor->take(*indices));
 }
 
-Tensor THSTensor_tan(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->tan());
-}
-
-Tensor THSTensor_tan_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->tan_());
-}
-
-Tensor THSTensor_tanh(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->tanh());
-}
-
-Tensor THSTensor_tanh_(const Tensor tensor)
-{
-    CATCH_TENSOR(tensor->tanh_());
-}
-
 Tensor THSTensor_tile(const Tensor tensor, const int64_t* rep, const int rep_length)
 {
     CATCH_TENSOR(tensor->tile(at::ArrayRef<int64_t>(rep, rep_length)));
@@ -2961,9 +1279,9 @@ void THSTensor_topk(const Tensor tensor, Tensor* (*allocator)(size_t length), co
 {
     CATCH(
         auto topk = tensor->topk(k, dim, largest, sorted);
-    Tensor * result = allocator(2);
-    result[0] = new torch::Tensor(std::get<0>(topk));
-    result[1] = new torch::Tensor(std::get<1>(topk));
+        Tensor * result = allocator(2);
+        result[0] = new torch::Tensor(std::get<0>(topk));
+        result[1] = new torch::Tensor(std::get<1>(topk));
     )
 }
 
@@ -2991,7 +1309,7 @@ Tensor THSTensor_to_device(const Tensor tensor, const int device_type, const int
 {
     CATCH_RETURN_Tensor(
         auto device = c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index);
-    res = ResultTensor(tensor->to(device));
+        res = ResultTensor(tensor->to(device)); 
     );
 }
 
@@ -3004,7 +1322,7 @@ Tensor THSTensor_to_type_and_device(const Tensor tensor, int8_t scalar_type, con
 {
     CATCH_RETURN_Tensor(
         auto device = c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index);
-    res = ResultTensor(tensor->to(device, at::ScalarType(scalar_type)));
+        res = ResultTensor(tensor->to(device, at::ScalarType(scalar_type)));
     );
 }
 
@@ -3044,89 +1362,6 @@ void THSTensor_unbind(const Tensor tensor, Tensor* (*allocator)(size_t length), 
     )
 }
 
-Tensor THSTensor_unsqueeze(Tensor tensor, int64_t dim)
-{
-    CATCH_TENSOR(tensor->unsqueeze(dim))
-}
-
-Tensor THSTensor_upsample_nearest1d(
-    const Tensor tensor,
-    const int64_t* outputSize, const int outputSizeLength,
-    const double* scaleFactors, const int scaleFactorsLength)
-{
-    CATCH_TENSOR(torch::upsample_nearest1d(
-        *tensor,
-        (outputSize == 0 ? c10::optional<c10::IntArrayRef>() : c10::optional<c10::IntArrayRef>(at::IntArrayRef(outputSize, outputSizeLength))),
-        (scaleFactors == 0 ? c10::optional<c10::ArrayRef<double>>() : c10::optional<c10::ArrayRef<double>>(at::ArrayRef<double>(scaleFactors, scaleFactorsLength))))
-    );
-}
-
-Tensor THSTensor_upsample_nearest1d_backward(
-    const Tensor grad_output,
-    const int64_t* outputSize, const int outputSizeLength,
-    const int64_t* inputSize, const int inputSizeLength,
-    const double* scaleFactors, const int scaleFactorsLength)
-{
-    CATCH_TENSOR(torch::upsample_nearest1d_backward(
-        *grad_output,
-        (outputSize == 0 ? c10::optional<c10::IntArrayRef>() : c10::optional<c10::IntArrayRef>(at::IntArrayRef(outputSize, outputSizeLength))),
-        at::IntArrayRef(inputSize, inputSizeLength),
-        (scaleFactors == 0 ? c10::optional<c10::ArrayRef<double>>() : c10::optional<c10::ArrayRef<double>>(at::ArrayRef<double>(scaleFactors, scaleFactorsLength))))
-    );
-}
-
-Tensor THSTensor_upsample_nearest2d(
-    const Tensor tensor,
-    const int64_t* outputSize, const int outputSizeLength,
-    const double* scaleFactors, const int scaleFactorsLength)
-{
-    CATCH_TENSOR(torch::upsample_nearest2d(
-        *tensor,
-        (outputSize == 0 ? c10::optional<c10::IntArrayRef>() : c10::optional<c10::IntArrayRef>(at::IntArrayRef(outputSize, outputSizeLength))),
-        (scaleFactors == 0 ? c10::optional<c10::ArrayRef<double>>() : c10::optional<c10::ArrayRef<double>>(at::ArrayRef<double>(scaleFactors, scaleFactorsLength))))
-    );
-}
-
-Tensor THSTensor_upsample_nearest2d_backward(
-    const Tensor grad_output,
-    const int64_t* outputSize, const int outputSizeLength,
-    const int64_t* inputSize, const int inputSizeLength,
-    const double* scaleFactors, const int scaleFactorsLength)
-{
-    CATCH_TENSOR(torch::upsample_nearest2d_backward(
-        *grad_output,
-        (outputSize == 0 ? c10::optional<c10::IntArrayRef>() : c10::optional<c10::IntArrayRef>(at::IntArrayRef(outputSize, outputSizeLength))),
-        at::IntArrayRef(inputSize, inputSizeLength),
-        (scaleFactors == 0 ? c10::optional<c10::ArrayRef<double>>() : c10::optional<c10::ArrayRef<double>>(at::ArrayRef<double>(scaleFactors, scaleFactorsLength))))
-    );
-}
-
-Tensor THSTensor_upsample_nearest3d(
-    const Tensor tensor,
-    const int64_t* outputSize, const int outputSizeLength,
-    const double* scaleFactors, const int scaleFactorsLength)
-{
-    CATCH_TENSOR(torch::upsample_nearest3d(
-        *tensor,
-        (outputSize == 0 ? c10::optional<c10::IntArrayRef>() : c10::optional<c10::IntArrayRef>(at::IntArrayRef(outputSize, outputSizeLength))),
-        (scaleFactors == 0 ? c10::optional<c10::ArrayRef<double>>() : c10::optional<c10::ArrayRef<double>>(at::ArrayRef<double>(scaleFactors, scaleFactorsLength))))
-    );
-}
-
-Tensor THSTensor_upsample_nearest3d_backward(
-    const Tensor grad_output,
-    const int64_t* outputSize, const int outputSizeLength,
-    const int64_t* inputSize, const int inputSizeLength,
-    const double* scaleFactors, const int scaleFactorsLength)
-{
-    CATCH_TENSOR(torch::upsample_nearest3d_backward(
-        *grad_output,
-        (outputSize == 0 ? c10::optional<c10::IntArrayRef>() : c10::optional<c10::IntArrayRef>(at::IntArrayRef(outputSize, outputSizeLength))),
-        at::IntArrayRef(inputSize, inputSizeLength),
-        (scaleFactors == 0 ? c10::optional<c10::ArrayRef<double>>() : c10::optional<c10::ArrayRef<double>>(at::ArrayRef<double>(scaleFactors, scaleFactorsLength))))
-    );
-}
-
 Tensor THSTensor_values(Tensor tensor)
 {
     CATCH_TENSOR(tensor->_values());
@@ -3135,85 +1370,6 @@ Tensor THSTensor_values(Tensor tensor)
 Tensor THSTensor_vander(const Tensor tensor, const int64_t N, const bool increasing)
 {
     CATCH_TENSOR(torch::vander(*tensor, N, increasing));
-}
-
-Tensor THSTensor_xlogy(const Tensor x, const Tensor y)
-{
-    CATCH_TENSOR(x->xlogy(*y));
-}
-
-Tensor THSTensor_zeros_out(const int64_t* sizes, const int length, const Tensor out)
-{
-    CATCH_TENSOR(torch::zeros_out(*out, at::ArrayRef<int64_t>(sizes, length)));
-}
-
-Tensor THSTensor_zeros(
-    const int64_t* sizes,
-    const int length,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::zeros(at::ArrayRef<int64_t>(sizes, length), options));
-}
-
-
-Tensor THSTensor_zeros_like(
-    const Tensor input,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::zeros_like(*input, options));
-}
-
-
-
-Tensor THSTensor_full_out(const int64_t* sizes, const int length, const Scalar value, const Tensor out)
-{
-    CATCH_TENSOR(torch::full_out(*out, at::ArrayRef<int64_t>(sizes, length), *value));
-}
-
-Tensor THSTensor_full(
-    const int64_t* sizes,
-    const int length,
-    const Scalar value,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::full(at::ArrayRef<int64_t>(sizes, length), *value, options));
-}
-
-
-Tensor THSTensor_full_like(
-    const Tensor input,
-    const Scalar value,
-    const int8_t scalar_type,
-    const int device_type, const int device_index,
-    const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::full_like(*input, *value, options));
 }
 
 // torch.lingalg:
@@ -3315,25 +1471,6 @@ Tensor THSLinalg_tensorsolve(const Tensor tensor, Tensor other, const int64_t* d
     CATCH_TENSOR(torch::linalg::tensorsolve(*tensor, *other, dims));
 }
 
-torch::nn::init::NonlinearityType get_nl_type(const int64_t nl)
-{
-    switch (nl)
-    {
-    default:
-    case 0:  return torch::kLinear;
-    case 1:  return torch::kConv1D;
-    case 2:  return torch::kConv2D;
-    case 3:  return torch::kConv3D;
-    case 4:  return torch::kConvTranspose1D;
-    case 5:  return torch::kConvTranspose2D;
-    case 6:  return torch::kConvTranspose3D;
-    case 7:  return torch::kSigmoid;
-    case 8:  return torch::kTanh;
-    case 9:  return torch::kReLU;
-    case 10: return torch::kLeakyReLU;
-    }
-}
-
 double THSInit_calculate_gain(int64_t nonlinearity, double param)
 {
     CATCH_RETURN(double, 0.0, torch::nn::init::calculate_gain(get_nl_type(nonlinearity), param))
@@ -3359,66 +1496,6 @@ Tensor THSInit_eye_(Tensor tensor)
     CATCH_TENSOR(torch::nn::init::eye_(*tensor))
 }
 
-Tensor THSTensor_bernoulli(const Tensor tensor, const double p, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->bernoulli(p) : tensor->bernoulli(p, *gen));
-}
-
-Tensor THSTensor_bernoulli_0(Tensor tensor, const double p, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->bernoulli_(p) : tensor->bernoulli_(p, *gen));
-}
-
-Tensor THSTensor_bernoulli_1(Tensor tensor, const Tensor p_tensor, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->bernoulli_(*p_tensor) : tensor->bernoulli_(*p_tensor, *gen));
-}
-
-Tensor THSTensor_multinomial(const Tensor tensor, const double num_samples, const bool replacement, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->multinomial(num_samples, replacement) : tensor->multinomial(num_samples, replacement, *gen));
-}
-
-Tensor THSTensor_cauchy_(Tensor tensor, const double median, const double sigma, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->cauchy_(median, sigma) : tensor->cauchy_(median, sigma, *gen));
-}
-
-Tensor THSTensor_exponential_(Tensor tensor, const double lambda, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->exponential_(lambda) : tensor->exponential_(lambda, *gen));
-}
-
-Tensor THSTensor_geometric_(Tensor tensor, const double p, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->geometric_(p) : tensor->geometric_(p, *gen));
-}
-
-Tensor THSTensor_log_normal_(Tensor tensor, double mean, double std, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->log_normal_(mean, std) : tensor->log_normal_(mean, std, *gen));
-}
-
-Tensor THSTensor_normal_(Tensor tensor, double mean, double std, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->normal_(mean, std) : tensor->normal_(mean, std, *gen));
-}
-
-Tensor THSTensor_random_(Tensor tensor, double low, double high, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->random_(low, high) : tensor->random_(low, high, *gen));
-}
-
-Tensor THSTensor_uniform_(Tensor tensor, double low, double high, const Generator gen)
-{
-    CATCH_TENSOR(gen == nullptr ? tensor->uniform_(low, high) : tensor->uniform_(low, high, *gen));
-}
-
-Tensor THSInit_normal_(Tensor tensor, double mean, double std)
-{
-    CATCH_TENSOR(torch::nn::init::normal_(*tensor, mean, std))
-}
-
 Tensor THSInit_ones_(Tensor tensor)
 {
     CATCH_TENSOR(torch::nn::init::ones_(*tensor))
@@ -3434,82 +1511,7 @@ Tensor THSInit_sparse_(Tensor tensor, double sparsity, double std)
     CATCH_TENSOR(torch::nn::init::sparse_(*tensor, sparsity, std))
 }
 
-Tensor THSInit_uniform_(Tensor tensor, double low, double high)
-{
-    CATCH_TENSOR(torch::nn::init::uniform_(*tensor, low, high))
-}
-
-Tensor THSInit_kaiming_normal_(Tensor tensor, double a, const int64_t mode, const int64_t nonlinearity)
-{
-    CATCH_TENSOR(torch::nn::init::kaiming_normal_(*tensor, a, mode == 0 ? torch::nn::init::FanModeType(torch::kFanIn) : torch::nn::init::FanModeType(torch::kFanOut), get_nl_type(nonlinearity)))
-}
-
-Tensor THSInit_kaiming_uniform_(Tensor tensor, double a, const int64_t mode, const int64_t nonlinearity)
-{
-    CATCH_TENSOR(torch::nn::init::kaiming_uniform_(*tensor, a, mode == 0 ? torch::nn::init::FanModeType(torch::kFanIn) : torch::nn::init::FanModeType(torch::kFanOut), get_nl_type(nonlinearity)))
-}
-
-Tensor THSInit_xavier_normal_(Tensor tensor, double gain)
-{
-    CATCH_TENSOR(torch::nn::init::xavier_normal_(*tensor, gain))
-}
-
-Tensor THSInit_xavier_uniform_(Tensor tensor, double gain)
-{
-    CATCH_TENSOR(torch::nn::init::xavier_uniform_(*tensor, gain))
-}
-
 Tensor THSInit_zeros_(Tensor tensor)
 {
     CATCH_TENSOR(torch::nn::init::zeros_(*tensor))
-}
-
-Tensor THSTensor_bartlett_window(const int64_t len, bool periodic, const int8_t scalar_type, const int device_type, const int device_index, const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::bartlett_window(len, periodic, options));
-}
-
-Tensor THSTensor_blackman_window(const int64_t len, bool periodic, const int8_t scalar_type, const int device_type, const int device_index, const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::blackman_window(len, periodic, options));
-}
-
-Tensor THSTensor_hamming_window(const int64_t len, bool periodic, double alpha, double beta, const int8_t scalar_type, const int device_type, const int device_index, const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::hamming_window(len, periodic, alpha, beta, options));
-}
-
-Tensor THSTensor_hann_window(const int64_t len, bool periodic, const int8_t scalar_type, const int device_type, const int device_index, const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::hann_window(len, periodic, options));
-}
-
-Tensor THSTensor_kaiser_window(const int64_t len, bool periodic, double beta, const int8_t scalar_type, const int device_type, const int device_index, const bool requires_grad)
-{
-    auto options = at::TensorOptions()
-        .dtype(at::ScalarType(scalar_type))
-        .device(c10::Device((c10::DeviceType)device_type, (c10::DeviceIndex)device_index))
-        .requires_grad(requires_grad);
-
-    CATCH_TENSOR(torch::kaiser_window(len, periodic, beta, options));
 }
